// src/modules/webhooks/dto/webhook.dto.ts
export class PaystackWebhookDto {
  event: string;
  data: {
    id: number;
    domain: string;
    status: string;
    reference: string;
    amount: number;
    message: string;
    gateway_response: string;
    paid_at: string;
    created_at: string;
    channel: string;
    currency: string;
    ip_address: string;
    metadata: Record<string, any>;
    customer: {
      id: number;
      first_name: string;
      last_name: string;
      email: string;
    };
  };
}